+++ src/actions/observation/ObservationChannelAction.js
@@ -1,0 +1,6 @@
+export function observationChannelAction(observation) {
+    return {
+        type: 'OBSERVATION_CHANNEL',
+        observation
+    };
+}
