+++ src/actions/websocket.js
@@ -1,13 +1,16 @@
 import * as SockJS from 'sockjs-client';
 import * as Stomp from 'stompjs';
 import { properties } from '../config/properties.js';
-import dispatcherHaProblemContent from './channel/PublishChannelContentAction';
-import dispatcherTransaction from './thankcoin/TransactionMediator';
 import { dispatcherChannel } from './channel/ChannelAction';
-import dispatcherObservation from "./observation/ObservationAction";
+import dispatcherObservation, {observationChannelAction} from "./observation/ObservationChannelAction";
 import { store } from '../App'
 import {dispatcherUserList} from "./user/UserAction";
 import {publishChannelContentIn} from "../door/PublishChannelContentDoor";
+import {transactionChannelIn} from "../door/TransactionChannelDoor";
+import {transactionChannelContentIn} from "../door/TransactionChannelContentDoor";
+import {observationChannelIn} from "../door/ObservationChannelDoor";
+import {getLoginUser} from "../components/common/LoginProcess";
+import {loginPromiseResolved} from "../components/common/ObservationProcess";
  var stompClient = null;
 export default function connect(username) {
     var socket = new SockJS(properties.notificationUrl);
@@ -18,10 +21,18 @@ export default function connect(username) {
             publishChannelContentIn(notification, store);
         });
 
-        stompClient.subscribe("/topic/sendThankCoin", function (notification) {
-            dispatcherTransaction(notification, store)
-        });
 
+       stompClient.subscribe("/topic/sendThankCoinChannel", function (data) {
+               var transaction=JSON.parse(data.body);
+
+
+           transactionChannelIn(transaction, store);
+        });
+       stompClient.subscribe("/topic/sendThankCoinChannelContent", function (data) {
+           var transaction=JSON.parse(data.body);
+
+           transactionChannelContentIn(transaction, store);
+         });
         stompClient.subscribe('/topic/pushNotificationChannel', function (notification) {
             dispatcherChannel(notification, store)
         });
@@ -33,7 +44,33 @@ export default function connect(username) {
 
         if(typeof username!=="undefined" && username.length>0){
             stompClient.subscribe("/user/" + username+ "/queue/sendObservationRequestSignal", function (notification) {
-                dispatcherObservation(notification, store)
+
+                var observation=JSON.parse(notification.body);
+                observation= observation.payload;
+
+                var objectType=observation.objectType;
+
+                var action;
+                var notificationMessage="";
+
+
+
+                switch (objectType) {
+                    case 'Channel':{
+
+                        action= observationChannelIn(observation);
+                        store.dispatch(action)
+                        break;
+                    }
+                }
+
+                getLoginUser().then( (loginUser) => loginPromiseResolved(loginUser, store, notificationMessage, observation))
+                    .catch(function(hata){
+
+                        console.log(hata)
+                    });
+
+
             });
         }
 
@@ -42,6 +79,8 @@ export default function connect(username) {
     });
 
 }
+
+
 export function isWebSocketConnected(isWebSocketConnected) {
     return {
         type: 'IS_WEBSOCKET_CONNECTED',
